backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0

# Supabase
supabase>=2.3.0

# AgentBay SDK
wuying-agentbay-sdk>=0.1.0

# OpenAI client (for model API)
openai>=1.12.0

# Image processing
Pillow>=10.2.0

# Environment variables
python-dotenv>=1.0.0

# Pydantic for data validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Phone Agent (local reference)
# Install with: pip install -e ../../

